services/base/spe-backend/docker/files/requirements.txt
alembic==1.10.2 
asyncio==3.4.3
opensearch-py==2.0.0
fastapi==0.95.0
gunicorn==20.1.0
httpx==0.22.0
minio==7.1.2
uvicorn==0.17.4
prometheus-api-client==0.4.2
psycopg2-binary==2.9.3
python-keycloak==2.9.0
python-multipart==0.0.5
pytest==7.0.0
requests-cache==1.0.1
SQLAlchemy==1.4.31
sqlalchemy-json==0.6.0
dicom_parser==1.2.3